Transitions work better/Still not perfect
Transitions are fixed so that the player doesn't get stuck any more but they are still a little weird. One collision block on the bottom left before transitioning to the right works. I am not sure why all the other blocks the player is able to slide through them, but I added more transitions so that if the player does slide into a block it will teleport them back to the middle where they should be.