Added 3 levels and a reset level button
Added 3 levels that are all the same but specify the name of the levels in the top left corner. Timer is effectively set to zero when player loses all three lives or wins beats level 3. As of right now beating level 3 sends you to level 1. FIXING: The score should be carried over from level to level and we need a title screen. Score needs to be reset when player dies but still keeping track of the score from the last level. Example: Level 1 you got 500 points then when you lose all your lives in level 2 the score should display 500 as the starting point.